Fix color export

+++ tools/convertX3D/type.h
@@ -69,7 +69,7 @@ typedef struct {
 
 typedef struct{
   uint32_t vertex_id[4]; //A,B,C,D
-  uint32_t RGB;
+  uint16_t RGB;
 } face;
 
 typedef struct{
